Refactor; changed Decimal to float

+++ apps/backend/src/modules/currency/core/entities/currency_to_calculate_entity.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -8,4 +7,4 @@ from modules.currency.core.enums.currency_enum import CurrencyEnum
 class CurrencyToCalculateEntity(BaseModel):
     currency_from: CurrencyEnum 
     currency_to: CurrencyEnum
-    value: Decimal
+    value: float
